+++ R/Shiny.R
@@ -14,6 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+.REQUIRED_OSM_VERION <- base::package_version("3.1.2")
+
 #' Launch the Diagnostics Explorer Shiny app
 #' @param connectionDetails An object of type \code{connectionDetails} as created using the
 #'                          \code{\link[DatabaseConnector]{createConnectionDetails}} function in the
@@ -115,6 +117,11 @@ launchDiagnosticsExplorer <- function(sqliteDbPath = "MergedCohortDiagnosticsDat
     } else {
       stop("Cannot continue without OhdsiShinyModulesPackage from github")
     }
+  }
+
+  osmVersion <- utils::packageVersion("OhdsiShinyModules")
+  if (osmVersion < .REQ_OSM_VERSION) {
+    cli::cli_warn("OhdsiShinyModules version {osmVersion} is out of date. It is suggested you update to at least {.REQUIRED_OSM_VERION}")
   }
 
   appDir <-